// src/app/features/members/models/member.model.ts
// member.model.ts

// export interface Member {
//   id?: any;
//   name: any;
//   age: any;
//   weight: any;
//   height: any;
// }

export interface Member {
  id?: any;
  name: any;
  age: any;
  weight: any;
  height: any;
  phoneNumber?: any;
  email?: any;
  goal?: any;
  leadSource?: any;
  status?: any;
}